pyproject.toml
[project]
name = "fake-it-till-you-make-it-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7",
    "SQLAlchemy>=2.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
]

[tool.uv]
# optional: if using uv as package manager
pip = { upgrade = true }
